// src/app/interfaces/grocery.interface.ts
export interface Grocery {
    name: Food;
    amount: number;
};

export enum Food {
    Steak = 'Steak',
    Chicken = 'Chicken',
    Pork = 'Pork',
    Rabbit = 'Rabbit',
    Crab = 'Crab',
    Shrimp = 'Shrimp',
    Tuna = 'Tuna',
    Salmon = 'Salmon',
    Cucumber = 'Cucumber',
    Onion = 'Onion',
    Corn = 'Corn',
    Potato = 'Potato',
    Broccoli = 'Broccoli',
    Tomato = 'Tomato',
    Apple = 'Apple',
    Orange = 'Orange',
    Pear = 'Pear',
    Pineapple = 'Pineapple',
    Lettuce = 'Lettuce',
    Bread = 'Bread',
    Roll = 'Roll',
    Cake = 'Cake',
    Pasta = 'Pasta',
    Baguette = 'Baguette',
    Cracker = 'Cracker',
    Salt = 'Salt',
    Pepper = 'Pepper',
    Garlic = 'Garlic',
    Oregano = 'Oregano',
    Basil = 'Basil',
    Paprika = 'Paprika',
    Cumin = 'Cumin',
    Milk = 'Milk',
    Cheddar = 'Cheddar',
    Pepperjack = 'Pepperjack',
    Swiss = 'Swiss',
    Yogurt = 'Yogurt',
    Parmesan = 'Parmesan',
    Mozzerella = 'Mozzerella'
 }

export enum Department {
    Butcher = 'Butcher',
    Produce = 'Produce',
    Bakery = 'Bakery',
    Seasonings = 'Seasonings',
    Dairy = 'Dairy'
}